DJANGO/projeto_escola/requirements.txt
altgraph==0.17
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
future==0.18.2
idna==2.10
Markdown==3.3.4
pefile==2019.4.18
Pillow==8.2.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.1.0
PyQt5-Qt==5.15.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
pyqt5-tools==5.15.4.3.0.3
PyRECONSTRUCT==2.2.0
python-dotenv==0.17.1
pytz==2021.1
pywin32-ctypes==0.2.0
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.4
